Support v2.6 and v2.7 simultaneously

+++ EEex-v2.6.6.0/COPY_TO_GAME.ps1
@@ -6,10 +6,10 @@ param (
 
 try {
 	if ($Env:DEPLOY -eq "1") {
-		"EEex.dll" | ForEach-Object {Copy-Item "$OutDir\$_" "$Env:DEPLOY_DIRECTORY\EEex\loader\$_"}
-		"EEex.pdb" | ForEach-Object {Copy-Item "$OutDir\$_" "$Env:DEPLOY_DIRECTORY\pdb\$_"}
+		"EEex.dll" | ForEach-Object {Copy-Item "$OutDir\$_" "$Env:DEPLOY_DIRECTORY\EEex\loader\v2.6.6.0\dll\$_"}
+		"EEex.pdb" | ForEach-Object {Copy-Item "$OutDir\$_" "$Env:DEPLOY_DIRECTORY\pdb\v2.6.6.0\$_"}
 	}
-	if (($GamePath -ne ".") -and (Test-Path $GamePath)) {
+	if (($Env:GAME_VERSION -eq "v2.6.6.0") -and ($GamePath -ne ".") -and (Test-Path $GamePath)) {
 		"EEex.dll", "EEex.pdb" | ForEach-Object {Copy-Item "$OutDir/$_" $GamePath}
 	}
 }
